refactor(home): split HomeView into focused modules
